refactor(platform-currencies): move test store wrapper to src/fixtures
Mirrors the domain api packages' fixtures/ convention: test-only support, not in the public barrel.
