+++ features/platform/currencies/src/hooks/useFeatureFlaggedCurrencies.test.tsx
@@ -1,6 +1,6 @@
 import { renderHook } from "@testing-library/react";
 import { useFeatureFlaggedCurrencies } from "./useFeatureFlaggedCurrencies";
-import { makeStoreWrapper, FEATURE_FLAGS_DEFAULTS } from "../renderWithStore";
+import { makeStoreWrapper, FEATURE_FLAGS_DEFAULTS } from "../fixtures";
 
 describe("useFeatureFlaggedCurrencies", () => {
   it("deactivates a currency whose gating flag is disabled by default", () => {
